llvm 3.8 compile fixes

+++ llarp/router.cpp
@@ -649,12 +649,12 @@ llarp_router::Run()
     InitServiceNode();
     // immediate connect all for service node
     auto delay = rand() % 100;
-    llarp_logic_call_later(logic, {delay, this, &ConnectAll});
+    llarp_logic_call_later(logic, {static_cast<uint64_t>(delay), this, &ConnectAll});
   }
   else
   {  // delayed connect all for clients
     auto delay = ((rand() % 10) * 500) + 1000;
-    llarp_logic_call_later(logic, {delay, this, &ConnectAll});
+    llarp_logic_call_later(logic, {static_cast<uint64_t>(delay), this, &ConnectAll});
   }
 
   llarp::PubKey ourPubkey = pubkey();
